Data.MessagePack

Description

Object is a message pack object, and it have constructors for all message pack types.

The Serialize instances define how Object values may be serialized and deserialized to message pack binary format, following the specification.

Documentation

data Object

Constructors

ObjectInt Int

ObjectNil

ObjectBool Bool

ObjectFloat Float

ObjectDouble Double

ObjectString Text

ObjectBinary ByteString

ObjectArray [Object]

ObjectMap (Map Object Object)